+++ lib/utils.php
@@ -1,0 +1,11 @@
+<?php
+    function verificaSession(){
+        session_start();
+        if($_SESSION && $_SESSION['user']){
+            return 1;
+        }else {
+            return 0;
+        }
+    }
+
+?>