+++ users/cadastra.php
@@ -1,4 +1,6 @@
 <?php
+include '../lib/utils.php';
+$login = verificaSession();
 ?>
 <!DOCTYPE html>
 <html lang="en">
@@ -16,6 +18,13 @@
     <header>
         <figure>
             <img src="" alt="logo">
+            <?php
+            if ($login !== 0) {
+                $name = $_SESSION['user']['nome'];
+                echo "<p>$name</p>";
+                echo '<a href="../lib/valida.php?logout">Logout</a>';
+            }
+            ?>
         </figure>
         <ul>
             <li> <a href="../">Home</a></li>