+++ changelog.md
@@ -1,8 +1,13 @@
 # Changelog
 
+## [v1.4.0 (12-07-2022)](https://github.com/wistech7l/sistema-concessionaria/releases/tag/v1.4.0)
+### Added
+* Adicionado Session de usuarios
+* Criação da página de acesso Restrito
+
 ## [v1.3.0 (12-07-2022)](https://github.com/wistech7l/sistema-concessionaria/releases/tag/v1.3.0)
 ### Added
-    * Criação do CRUD para usuários
+* Criação do CRUD para usuários
 
 ## [v1.2.0 (06-07-2022)](https://github.com/wistech7l/sistema-concessionaria/releases/tag/v1.2.0)
 ### Added